requirements.txt
# FastAPI и ASGI сервер
fastapi==0.104.1
uvicorn==0.24.0

# База данных
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Безопасность и аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Обработка переменных окружения
python-dotenv==1.0.0

# Обработка CORS
starlette==0.27.0

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0

# Работа с датами
python-dateutil==2.8.2

# Для работы с файлами
aiofiles==23.2.1

# Для тестирования
pytest==7.4.3
httpx==0.25.2

# Для документации
mkdocs==1.5.3
mkdocs-material==9.5.13 